GEN_AI/resume-screening-ai/requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
openai
python-dotenv
